+++ duckscript_sdk/src/sdk/std/collections/mod.rs
@@ -6,6 +6,7 @@ pub(crate) mod array_length;
 pub(crate) mod array_pop;
 mod array_push;
 mod is_array;
+mod is_map;
 mod map;
 mod map_get;
 mod map_put;
@@ -30,6 +31,7 @@ pub(crate) fn load(commands: &mut Commands, parent: &str) -> Result<(), ScriptEr
     commands.set(array_length::create(&package))?;
     commands.set(array_pop::create(&package))?;
     commands.set(is_array::create(&package))?;
+    commands.set(is_map::create(&package))?;
     commands.set(map::create(&package))?;
     commands.set(map_get::create(&package))?;
     commands.set(map_put::create(&package))?;